use long long instead of long (#5)

+++ src/domain/timetoken.cpp
@@ -5,7 +5,7 @@ Timetoken::Timetoken Timetoken::now() {
     return std::to_string(now_numeric());
 }
 
-long Timetoken::now_numeric() {
+long long Timetoken::now_numeric() {
     // TODO: move time to infra for testing purposes
     return std::chrono::duration_cast<std::chrono::nanoseconds>(
             std::chrono::system_clock::now().time_since_epoch()).count();
@@ -19,6 +19,6 @@ Timetoken::Timetoken Timetoken::increase_by_one(const Timetoken& timetoken) {
     return increase_by(timetoken, 1);
 }
 
-long Timetoken::to_long(const Timetoken& timetoken) {
-    return std::stol(timetoken);
+long long Timetoken::to_long(const Timetoken& timetoken) {
+    return std::stoll(timetoken);
 }
